+++ pkgs/test/test/runner/line_and_col_test.dart
@@ -56,13 +56,49 @@ void main() {
       await test.shouldExit(0);
     });
 
+    test('additionally selects test with matching custom location', () async {
+      await d.dir('test').create();
+      var testFileUri = Uri.file(d.file('test/aaa_test.dart').io.path);
+      var notTestFileUri = Uri.file(d.file('test/bbb.dart').io.path);
+      var testFilePackageRelativeUri =
+          Uri.parse('org-dartlang-app:///test/aaa_test.dart');
+      await d.file('test/aaa_test.dart', '''
+        import 'package:test/test.dart';
+
+        void main() {
+          test("a", () {}); // Line 4 from stack trace (match)
+
+          // Custom line 4 (match)
+          test("b", location: TestLocation(Uri.parse('$testFileUri'), 4, 0), () {});
+
+          // Custom line 4 match but using org-dartlang-app (match)
+          test("c", location: TestLocation(Uri.parse('$testFilePackageRelativeUri'), 4, 0), () {});
+
+          // Custom different line, same file (no match)
+          test("d", location: TestLocation(Uri.parse('$testFileUri'), 5, 0), () => throw TestFailure("oh no"));
+
+          // Custom line 4 match but different file (no match)
+          test("e", location: TestLocation(Uri.parse('$notTestFileUri'), 4, 0), () => throw TestFailure("oh no"));
+        }
+      ''').create();
+
+      var test = await runTest(['test/aaa_test.dart?line=4']);
+
+      expect(
+        test.stdout,
+        emitsThrough(contains('+3: All tests passed!')),
+      );
+
+      await test.shouldExit(0);
+    });
+
     test('selects groups with a matching line', () async {
       await d.file('test.dart', '''
         import 'package:test/test.dart';
 
         void main() {
           group("a", () {
-            test("b", () {});
+            test("a", () {});
           });
           group("b", () {
             test("b", () => throw TestFailure("oh no"));
@@ -75,6 +111,53 @@ void main() {
       expect(
         test.stdout,
         emitsThrough(contains('+1: All tests passed!')),
+      );
+
+      await test.shouldExit(0);
+    });
+
+    test('additionally selects groups with a matching custom location',
+        () async {
+      await d.dir('test').create();
+      var testFileUri = Uri.file(d.file('test/aaa_test.dart').io.path);
+      var notTestFileUri = Uri.file(d.file('test/bbb.dart').io.path);
+      var testFilePackageRelativeUri =
+          Uri.parse('org-dartlang-app:///test/aaa_test.dart');
+      await d.file('test/aaa_test.dart', '''
+        import 'package:test/test.dart';
+
+        void main() {
+          group("a", () { // Line 4 from stack trace (match)
+            test("a", () {});
+          });
+
+          // Custom line 4 (match)
+          group("b", location: TestLocation(Uri.parse('$testFileUri'), 4, 0), () {
+            test("b", () {});
+          });
+
+          // Custom line 4 match but using org-dartlang-app (match)
+          group("c", location: TestLocation(Uri.parse('$testFilePackageRelativeUri'), 4, 0), () {
+            test("c", () {});
+          });
+
+          // Custom different line, same file (no match)
+          group("d", location: TestLocation(Uri.parse('$testFileUri'), 5, 0), () {
+            test("d", () => throw TestFailure("oh no"));
+          });
+
+          // Custom line 4 match but different file (no match)
+          group("e", location: TestLocation(Uri.parse('$notTestFileUri'), 4, 0), () {
+            test("e", () => throw TestFailure("oh no"));
+          });
+        }
+      ''').create();
+
+      var test = await runTest(['test/aaa_test.dart?line=4']);
+
+      expect(
+        test.stdout,
+        emitsThrough(contains('+3: All tests passed!')),
       );
 
       await test.shouldExit(0);