+++ pkgs/test_core/lib/src/runner.dart
@@ -309,11 +309,12 @@ class Runner {
             var line = selection.line;
             var col = selection.col;
             if (line == null && col == null) return true;
+
             var trace = test.trace;
-            if (trace == null) {
+            if (trace == null && test.location == null) {
               throw StateError(
                   'Cannot filter by line/column for this test suite, no stack'
-                  'trace available.');
+                  'trace or location available.');
             }
             var path = suite.path;
             if (path == null) {
@@ -322,12 +323,14 @@ class Runner {
                   'path available.');
             }
             // The absolute path as it will appear in stack traces.
-            var absoluteSuitePath = File(path).absolute.uri.toFilePath();
+            var suiteUri = File(path).absolute.uri;
+            var absoluteSuitePath = suiteUri.toFilePath();
 
-            bool matchLineAndCol(Frame frame) {
-              switch (frame.uri.scheme) {
+            /// Helper to check if [uri] matches the suite path.
+            bool matchesUri(Uri uri) {
+              switch (uri.scheme) {
                 case 'file':
-                  if (frame.uri.toFilePath() != absoluteSuitePath) {
+                  if (uri.toFilePath() != absoluteSuitePath) {
                     return false;
                   }
                 case 'package':
@@ -338,10 +341,33 @@ class Runner {
                   // Now we can assume that the kernel is compiled using
                   // --filesystem-scheme.
                   // In this case, because we don't know the --filesystem-root, as
-                  // long as the file path matches we assume it is the same file.
-                  if (!absoluteSuitePath.endsWith(frame.uri.path)) {
+                  // long as the path matches we assume it is the same file.
+                  if (!suiteUri.path.endsWith(uri.path)) {
                     return false;
                   }
+              }
+
+              return true;
+            }
+
+            // First check if we're a match for the overridden location for this
+            // item or any parents.
+            var current = test as GroupEntry?;
+            while (current != null) {
+              if (current.location case var location?) {
+                if ((line == null || location.line == line) &&
+                    (col == null || location.column == col) &&
+                    matchesUri(location.uri)) {
+                  return true;
+                }
+              }
+              current = current.parent;
+            }
+
+            /// Helper to check if [frame] matches the suite path, line and col.
+            bool matchLineAndCol(Frame frame) {
+              if (!matchesUri(frame.uri)) {
+                return false;
               }
               if (line != null && frame.line != line) {
                 return false;
@@ -352,7 +378,8 @@ class Runner {
               return true;
             }
 
-            return trace.frames.any(matchLineAndCol);
+            // Check if any frames in the stack trace match.
+            return trace?.frames.any(matchLineAndCol) ?? false;
           });
         }));
       });