+++ pkgs/test_core/lib/src/scaffolding.dart
@@ -140,6 +140,7 @@ void test(Object? description, dynamic Function() body,
     Object? tags,
     Map<String, dynamic>? onPlatform,
     int? retry,
+    TestLocation? location,
     // TODO(https://github.com/dart-lang/test/issues/2205): Remove deprecated.
     @Deprecated('Debug only') @doNotSubmit bool solo = false}) {
   _declarer.test(description.toString(), body,
@@ -149,6 +150,7 @@ void test(Object? description, dynamic Function() body,
       onPlatform: onPlatform,
       tags: tags,
       retry: retry,
+      location: location,
       solo: solo);
 
   // Force dart2js not to inline this function. We need it to be separate from
@@ -219,6 +221,7 @@ void group(Object? description, dynamic Function() body,
     Object? tags,
     Map<String, dynamic>? onPlatform,
     int? retry,
+    TestLocation? location,
     // TODO(https://github.com/dart-lang/test/issues/2205): Remove deprecated.
     @Deprecated('Debug only') @doNotSubmit bool solo = false}) {
   _declarer.group(description.toString(), body,
@@ -228,6 +231,7 @@ void group(Object? description, dynamic Function() body,
       tags: tags,
       onPlatform: onPlatform,
       retry: retry,
+      location: location,
       solo: solo);
 
   // Force dart2js not to inline this function. We need it to be separate from
@@ -278,7 +282,8 @@ void tearDown(dynamic Function() callback) => _declarer.tearDown(callback);
 /// dependencies between tests that should be isolated. In general, you should
 /// prefer [setUp], and only use [setUpAll] if the callback is prohibitively
 /// slow.
-void setUpAll(dynamic Function() callback) => _declarer.setUpAll(callback);
+void setUpAll(dynamic Function() callback, {TestLocation? location}) =>
+    _declarer.setUpAll(callback, location: location);
 
 /// Registers a function to be run once after all tests.
 ///
@@ -291,5 +296,5 @@ void setUpAll(dynamic Function() callback) => _declarer.setUpAll(callback);
 /// dependencies between tests that should be isolated. In general, you should
 /// prefer [tearDown], and only use [tearDownAll] if the callback is
 /// prohibitively slow.
-void tearDownAll(dynamic Function() callback) =>
-    _declarer.tearDownAll(callback);
+void tearDownAll(dynamic Function() callback, {TestLocation? location}) =>
+    _declarer.tearDownAll(callback, location: location);