+++ pkgs/test_core/CHANGELOG.md
@@ -1,6 +1,9 @@
 ## 0.6.9-wip
 
-- Add support for native assets for `dart test` in pub workspaces.
+* Add support for native assets for `dart test` in pub workspaces.
+* `test()` and `group()` functions now take an optional `TestLocation` that will
+  be used as the location of the test in JSON reporters instead of being parsed
+  from the call stack.
 
 ## 0.6.8