+++ pkgs/test_core/CHANGELOG.md
@@ -4,6 +4,7 @@
 * `test()` and `group()` functions now take an optional `TestLocation` that will
   be used as the location of the test in JSON reporters instead of being parsed
   from the call stack.
+* Set a debug name for test isolates.
 
 ## 0.6.8